Move some stuff around and add delete webhook method
